requirements.txt
langchain
openai
pinecone-client
python-dotenv
tiktoken
pypdf